rework MemoStack back into a separate type
https://github.com/open-policy-agent/swift-opa/pull/59 was a bit
aggressive in flattening the MemoStack functionality into the
IREvaluationContext. This led to unclear division of responsibility and
APIs - for example having the subscript operator of IREvaluationContext
dereference the memo stack is surprising.

This commit works part of that back by separating the MemoStack into its
previously distinct type. This should not impact the performance gains
of #59, as a lot of that was likely due to removal of the Frame
functionality.
